Complete HTML project for TOP
That is the first completed project for the Odin Project.
It's about the recipes and steps how to cook as well as ingredients to be needed.
With images of that ready dishes, it can give give a sense of a real recipe website.
Although, it is the new and amateour one, but I'm on my way :)